requirements.txt
pillow <= 10.4.0
pandas >= 2.2.3
numpy <= 1.26.0
matplotlib
tqdm
parmed
rdkit >= 2024.03.5
seaborn
alchemlyb
parse
scipy <= 1.12.0
scikit-learn
pymbar <= 3.1.1
alive-progress
openmmtools
openmm-plumed
openpyxl
toml
reportlab
pypdf2 > 3.0.0
jinja2
pdb2pqr
pygraphviz